Downloaded develop branch from https://nuserv.uni-muenster.de:8443/katrin-git/kassiopeia for purpose of reading arbitrary field map.

+++ KEMField/Source/Applications/Test/TestBinaryTruncation.cc
@@ -105,7 +105,7 @@ double Round(double x, unsigned int masked)
     if (maskedBits)
     {
       index = endian_min + endian_dir*maskedBytes;
-      significantBit = (~(0xff << 1)) << (maskedBits - 1);
+      significantBit = (unsigned short)(~(0xff << 1)) << (maskedBits - 1);
     }
     else
     {
